third_party/blink/web_tests/fast/filesystem/filesystem-reference-expected.txt

This tests if Entries returned by callbacks keep functioning even after we lose a reference to the filesystem in the script.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Running runTest1()...
Running runTest2()...
Running runTest3()...
Running runTest4()...
PASS testEntries.length is 4
PASS successfullyParsed is true

TEST COMPLETE
